Final commit for the day.
Changed around some positions in the scene file
Added timers that display on the screen! This is so that users can see if they're selecting something and for how long.

Every single Canvas is now activated and de-activated when needed along with the scripts.

Commented all of the code in my scripts so that they can be human-read now.